+++ src/database/data-source.ts
@@ -4,6 +4,7 @@ import { User } from '../users/entities/user.entity';
 import { Customer } from '../customers/entities/customer.entity';
 import { FlussonicServer } from '../flussonic-servers/entities/flussonic-server.entity';
 import { FlussonicServerStat } from '../flussonic-servers/entities/flussonic-server-stat.entity';
+import { FlussonicStream } from '../flussonic-servers/entities/flussonic-stream.entity';
 
 config();
 
@@ -14,7 +15,13 @@ export const dataSourceOptions: DataSourceOptions = {
   username: process.env.DB_USERNAME ?? 'root',
   password: process.env.DB_PASSWORD ?? '',
   database: process.env.DB_DATABASE ?? 'project7_auth',
-  entities: [User, Customer, FlussonicServer, FlussonicServerStat],
+  entities: [
+    User,
+    Customer,
+    FlussonicServer,
+    FlussonicServerStat,
+    FlussonicStream,
+  ],
   migrations: [__dirname + '/migrations/*.{ts,js}'],
   synchronize: false,
   logging: process.env.APP_ENV === 'development',